refac: add Coin type in a separated file
