+++ client/src/types/index.ts
@@ -1,0 +1,6 @@
+export interface Coin {
+  assetId: string;
+  symbol?: string;
+  name?: string;
+  img?: string;
+}
